requirements.txt
bert-extractive-summarizer==0.7.1
spacy==2.0.12
transformers==2.2.0
pandas
torch